chat: don't duplicate modified entries in every edit stop
Now that we have `ensureEditInUndoStopMatches`, we don't need to
aggressively store every file in every snapshot. In this change, the
file is only stored at the initial complete snapshot for each request,
and then in each edit stop in which it changes. Only requires a slight
change in getFirstAndLastStop.

Also popped storage out to its own file and wrote some sanity tests for
it, although I didn't find much low hanging fruit there.

Fixes https://github.com/microsoft/vscode-copilot/issues/14687
